Updating the code to connect to the postgres database.
The previous code was the result of diffent copy/paste
to understand how things are working.
This commit structures the code in small functions to get
some readable code. From this base, it should be easier to
extract code to get reusable stuff.
I also added some error handling using the library from
Dave Cheny, github.com/pkg/errors .
At last, I have externalized the configuration to connect to
the database so we don't have to recompile the code to
change the database to talk to.
The configuration file is part of this commit.